Replace long() with int()
long() is deprecated and merged with int.

+++ lib/python/Screens/InfoBarGenerics.py
@@ -3168,7 +3168,7 @@ class InfoBarCueSheetSupport:
 		r = seek.getPlayPosition()
 		if r[0]:
 			return None
-		return long(r[1])
+		return int(r[1])
 
 	def cueGetEndCutPosition(self):
 		ret = False
