+++ lib/python/Tools/CIHelper.py
@@ -43,7 +43,7 @@ class CIHelper:
 
 						for caid in slot.findall("caid"):
 							read_caid = caid.get("id").encode("UTF-8")
-							usingcaid.append(long(read_caid, 16))
+							usingcaid.append(int(read_caid, 16))
 
 						for service in slot.findall("service"):
 							read_service_ref = service.get("ref").encode("UTF-8")
@@ -54,7 +54,7 @@ class CIHelper:
 						for provider in slot.findall("provider"):
 							read_provider_name = provider.get("name").encode("UTF-8")
 							read_provider_dvbname = provider.get("dvbnamespace").encode("UTF-8")
-							read_providers.append((read_provider_name, long(read_provider_dvbname, 16)))
+							read_providers.append((read_provider_name, int(read_provider_dvbname, 16)))
 							if read_slot is not False:
 								provider_services_refs = self.getProivderServices([read_provider_name])
 								if provider_services_refs:
